Allow users to generate their own grids
Added a button that gives users the ability to generate their own grid.
When clicking the button at the top of the screen, a grid will be
generated using the number inputed by the user to set the amount of
rows and columns. When clicking the button again after a grid has
already been generated, the previous grid will be cleared and a new grid
will be generated. The max value for users to input is 100. Any value
higher than that will default to 100.